Reduce code repetition
